Renames useWindowResizer file
