+++ src/forms/TopComponent/TopComponent.test.tsx
@@ -1,5 +1,5 @@
 import { ButtonProps } from "@root/components/Button";
-import { ViewProvider } from "@root/utils/useWindowResizer";
+import { ViewProvider } from "@root/utils/formViewUtils";
 import { render, screen, cleanup, fireEvent } from "@testing-library/react";
 import * as React from "react";
